requirements.txt
# Core Framework
flask==2.3.2
python-dotenv==1.0.0

# Machine Learning / Deep Learning
tensorflow==2.13.0
scikit-learn==1.2.2
joblib==1.3.2

# Spotify API
spotipy==2.23.0

# Utilities
pandas==2.0.3
numpy==1.24.3
